added blueshift.js
